Move notification and notification jobs to the core
