+++ src/FunderMaps.BatchNode/Startup.cs
@@ -39,8 +39,6 @@ namespace FunderMaps.BatchNode
             // Add batch jobs to the DI container.
             services.AddBatchJob<Jobs.BundleBuilder.BundleBatch>();
             services.AddBatchJob<Jobs.BundleBuilder.BundleJob>();
-            services.AddBatchJob<Jobs.Notification.EmailJob>();
-            services.AddBatchJob<Jobs.IncidentNotifyJob>();
 
             services.AddGrpc();
