+++ src/FunderMaps.Core/IncidentReport/IncidentService.cs
@@ -3,8 +3,10 @@ using FunderMaps.Core.Interfaces;
 using FunderMaps.Core.Interfaces.Repositories;
 using FunderMaps.Core.Notification;
 using Microsoft.Extensions.Options;
+using Scriban.Runtime;
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 
 namespace FunderMaps.Core.IncidentReport
@@ -16,6 +18,7 @@ namespace FunderMaps.Core.IncidentReport
     {
         private readonly IncidentOptions _options;
         private readonly Core.AppContext _appContext;
+        private readonly ITemplateParser _templateParser;
         private readonly IContactRepository _contactRepository;
         private readonly IIncidentRepository _incidentRepository;
         private readonly IGeocoderTranslation _geocoderTranslation;
@@ -27,17 +30,105 @@ namespace FunderMaps.Core.IncidentReport
         public IncidentService(
             IOptions<IncidentOptions> options,
             Core.AppContext appContext,
+            ITemplateParser templateParser,
             IContactRepository contactRepository,
             IIncidentRepository incidentRepository,
             IGeocoderTranslation geocoderTranslation,
-            INotifyService notifyService)
+            INotifyService notificationService)
         {
             _options = options?.Value ?? throw new ArgumentNullException(nameof(options));
             _appContext = appContext ?? throw new ArgumentNullException(nameof(appContext));
+            _templateParser = templateParser ?? throw new ArgumentNullException(nameof(templateParser));
             _contactRepository = contactRepository ?? throw new ArgumentNullException(nameof(contactRepository));
             _incidentRepository = incidentRepository ?? throw new ArgumentNullException(nameof(incidentRepository));
             _geocoderTranslation = geocoderTranslation ?? throw new ArgumentNullException(nameof(geocoderTranslation));
-            _notifyService = notifyService ?? throw new ArgumentNullException(nameof(notifyService));
+            _notifyService = notificationService ?? throw new ArgumentNullException(nameof(notificationService));
+        }
+
+        // TODO: This is a temporary solution.
+        // TODO: Move into culture service?
+        class TranslationFunctions : ScriptObject
+        {
+            public static string ToBoolean(bool value) => value ? "Ja" : "Nee";
+
+            public static string ToFoundationType(Core.Types.FoundationType value)
+                => value switch
+                {
+                    Core.Types.FoundationType.Wood => "Hout",
+                    Core.Types.FoundationType.WoodAmsterdam => "Hout",
+                    Core.Types.FoundationType.WoodRotterdam => "Hout",
+                    Core.Types.FoundationType.WoodCharger => "Hout",
+                    Core.Types.FoundationType.Concrete => "Beton",
+                    Core.Types.FoundationType.NoPile => "Niet onderheid",
+                    Core.Types.FoundationType.NoPileMasonry => "Niet onderheid",
+                    Core.Types.FoundationType.NoPileStrips => "Niet onderheid",
+                    Core.Types.FoundationType.NoPileBearingFloor => "Niet onderheid",
+                    Core.Types.FoundationType.NoPileConcreteFloor => "Niet onderheid",
+                    Core.Types.FoundationType.NoPileSlit => "Niet onderheid",
+                    Core.Types.FoundationType.WeightedPile => "Beton met verzwaardepunt",
+                    Core.Types.FoundationType.Combined => "Gecombineerd",
+                    Core.Types.FoundationType.SteelPile => "Stalen buispaal",
+                    Core.Types.FoundationType.Other => "Overig",
+                    _ => "Onbekend",
+                };
+
+            public static string ToFoundationDamageCause(Core.Types.FoundationDamageCause value)
+                => value switch
+                {
+                    Core.Types.FoundationDamageCause.Drainage => "Ontwateringsdiepte onvoldoende",
+                    Core.Types.FoundationDamageCause.Drystand => "Droogstand",
+                    Core.Types.FoundationDamageCause.ConstructionFlaw => "Verkeerd gefundeerd",
+                    Core.Types.FoundationDamageCause.Overcharge => "Overbelasting",
+                    Core.Types.FoundationDamageCause.OverchargeNegativeCling => "Overbelasting/Negatievekleef",
+                    Core.Types.FoundationDamageCause.NegativeCling => "Negatievekleef",
+                    Core.Types.FoundationDamageCause.BioInfection => "Bacteriele aantasting",
+                    Core.Types.FoundationDamageCause.FungusInfection => "Schimmelaantasting",
+                    Core.Types.FoundationDamageCause.BioFungusInfection => "Schimmel/bacterieen",
+                    Core.Types.FoundationDamageCause.FoundationFlaw => "Verkeerd gefundeerd",
+                    Core.Types.FoundationDamageCause.ConstructionHeave => "Woning omhooggedrukt",
+                    Core.Types.FoundationDamageCause.Subsidence => "Bodemdaling",
+                    Core.Types.FoundationDamageCause.Vegetation => "Wortels/planten",
+                    Core.Types.FoundationDamageCause.Gas => "Gaswinning/mijnbouw",
+                    Core.Types.FoundationDamageCause.Vibrations => "Verkeer",
+                    Core.Types.FoundationDamageCause.PartialFoundationRecovery => "Naastgelegen funderingsherstel",
+                    _ => "Onbekend",
+                };
+
+            public static string ToFoundationDamageCharacteristics(Core.Types.FoundationDamageCharacteristics value)
+                => value switch
+                {
+                    Core.Types.FoundationDamageCharacteristics.JammingDoorWindow => "Klemmende ramen/deuren",
+                    Core.Types.FoundationDamageCharacteristics.Crack => "Scheuren",
+                    Core.Types.FoundationDamageCharacteristics.Skewed => "Scheefstand",
+                    Core.Types.FoundationDamageCharacteristics.CrawlspaceFlooding => "Water in kruipruimte",
+                    Core.Types.FoundationDamageCharacteristics.ThresholdAboveSubsurface => "Maaiveld lager dan dorpel",
+                    Core.Types.FoundationDamageCharacteristics.ThresholdBelowSubsurface => "Drempel lager dan maaiveld",
+                    Core.Types.FoundationDamageCharacteristics.CrookedFloorWall => "Scheve vloer",
+                    _ => "Onbekend",
+                };
+
+            public static IEnumerable<string> ArrayToFoundationDamageCharacteristics(IEnumerable<Core.Types.FoundationDamageCharacteristics> values)
+                => values.Select(value => ToFoundationDamageCharacteristics(value));
+
+            public static string ToEnvironmentDamageCharacteristics(Core.Types.EnvironmentDamageCharacteristics value)
+                => value switch
+                {
+                    Core.Types.EnvironmentDamageCharacteristics.Subsidence => "Bodemdaling",
+                    Core.Types.EnvironmentDamageCharacteristics.SaggingSewerConnection => "Verzakkend riool",
+                    Core.Types.EnvironmentDamageCharacteristics.SaggingCablesPipes => "Verzakkende kabels/leidingen",
+                    Core.Types.EnvironmentDamageCharacteristics.Flooding => "Wateroverlast",
+                    Core.Types.EnvironmentDamageCharacteristics.FoundationDamageNearby => "Funderingschade in wijk",
+                    Core.Types.EnvironmentDamageCharacteristics.Elevation => "Recent opgehoogd",
+                    Core.Types.EnvironmentDamageCharacteristics.IncreasingTraffic => "Verkeerstoename",
+                    Core.Types.EnvironmentDamageCharacteristics.ConstructionNearby => "Werkzaamheden in wijk",
+                    Core.Types.EnvironmentDamageCharacteristics.VegetationNearby => "Bomen nabij",
+                    Core.Types.EnvironmentDamageCharacteristics.SewageLeakage => "Lekkend riool",
+                    Core.Types.EnvironmentDamageCharacteristics.LowGroundWater => "Wateronderlast",
+                    _ => "Onbekend",
+                };
+
+            public static IEnumerable<string> ArrayToEnvironmentDamageCharacteristics(IEnumerable<Core.Types.EnvironmentDamageCharacteristics> values)
+                => values.Select(value => ToEnvironmentDamageCharacteristics(value));
         }
 
         // FUTURE: split logic, hard to read.
@@ -55,14 +146,23 @@ namespace FunderMaps.Core.IncidentReport
             incident.Meta = meta;
 
             await _contactRepository.AddAsync(incident.ContactNavigation);
-            var id = await _incidentRepository.AddAsync(incident);
-            incident = await _incidentRepository.GetByIdAsync(id);
-            incident.ContactNavigation = await _contactRepository.GetByIdAsync(incident.Email);
+            incident = await _incidentRepository.AddGetAsync(incident);
+            var contact = await _contactRepository.GetByIdAsync(incident.Email);
+            incident.ContactNavigation = contact;
 
-            await _notifyService.DispatchNotifyAsync("incident_notify", new()
+            _templateParser.AddObject(nameof(incident), incident);
+            _templateParser.AddObject(nameof(address), address);
+            _templateParser.AddObject(nameof(contact), contact);
+            _templateParser.FromTemplateFile("Email", "incident");
+
+            // FUTURE: Fix this cast.
+            (_templateParser as Core.Components.TemplateParser).RegisterExtension(new TranslationFunctions());
+
+            await _notifyService.DispatchNotifyAsync(new()
             {
                 Recipients = _options.Recipients,
-                Items = new Dictionary<string, string> { { "id", id } },
+                Content = await _templateParser.RenderAsync(_appContext.CancellationToken),
+                Subject = $"Nieuwe melding: {incident.Id}",
             });
 
             return incident;
