+++ src/FunderMaps.Core/Notification/NotificationHub.cs
@@ -18,10 +18,7 @@ namespace FunderMaps.Core.Notification
         ///     Create new instance.
         /// </summary>
         public NotificationHub(AppContext appContext, BackgroundTaskScopedDispatcher backgroundTaskDispatcher)
-        {
-            AppContext = appContext;
-            _backgroundTaskDispatcher = backgroundTaskDispatcher;
-        }
+            => (AppContext, _backgroundTaskDispatcher) = (appContext, backgroundTaskDispatcher);
 
         /// <summary>
         ///     Notify by means of contacting.
@@ -29,14 +26,6 @@ namespace FunderMaps.Core.Notification
         /// <param name="envelope">Envelope containing the notification.</param>
         public async Task DispatchNotifyAsync(Envelope envelope)
             => await _backgroundTaskDispatcher.EnqueueTaskAsync(TaskName, envelope);
-
-        /// <summary>
-        ///     Notify by means of contacting.
-        /// </summary>
-        /// <param name="taskName">Name of the task to handle the job.</param>
-        /// <param name="envelope">Envelope containing the notification.</param>
-        public async Task DispatchNotifyAsync(string taskName, Envelope envelope)
-            => await _backgroundTaskDispatcher.EnqueueTaskAsync(taskName, envelope);
     }
 }
 #pragma warning restore CA1812 // Internal class is never instantiated
